+++ app/src/main/java/com/app/arcabyolimpo/data/repository/disabilities/DisabilityRepositoryImpl.kt
@@ -2,6 +2,7 @@ package com.app.arcabyolimpo.data.repository.disabilities
 
 import android.util.Log
 import com.app.arcabyolimpo.data.mapper.disabilities.toRegisterDto
+import com.app.arcabyolimpo.data.mapper.disabilities.toDomain
 import com.app.arcabyolimpo.data.remote.api.ArcaApi
 import com.app.arcabyolimpo.domain.model.disabilities.Disability
 import com.app.arcabyolimpo.domain.repository.disability.DisabilityRepository
@@ -54,4 +55,14 @@ class DisabilityRepositoryImpl
             val dto = disability.toRegisterDto()
             api.registerDisability(dto)
         }
+
+        /**
+         * Retrieves the details of the selected disability from the API.
+         *
+         * The response is converted into a [Disability] domain model.
+         *
+         * @return A [Disabilities] object representing the available disabilities.
+         */
+        override suspend fun getDisability(id: String): Disability = api.getDisabilityDetail(id).toDomain()
     }
+