+++ app/src/main/java/com/app/arcabyolimpo/domain/repository/beneficiaries/BeneficiaryRepository.kt
@@ -18,4 +18,5 @@ interface BeneficiaryRepository {
     suspend fun searchBeneficiaries(query: String): List<Beneficiary>
 
     suspend fun addBeneficiary(newBeneficiary: BeneficiaryDto): Beneficiary
+    suspend fun modifyBeneficiary(modifiedBeneficiary: BeneficiaryDto): Beneficiary
 }