+++ app/src/main/java/com/app/arcabyolimpo/presentation/navigation/NavGraph.kt
@@ -1,5 +1,6 @@
 package com.app.arcabyolimpo.presentation.navigation
 
+import android.util.Log
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
@@ -18,12 +19,14 @@ import com.app.arcabyolimpo.data.remote.interceptor.SessionManager
 import com.app.arcabyolimpo.domain.model.auth.UserRole
 import com.app.arcabyolimpo.presentation.common.components.LoadingShimmer
 import com.app.arcabyolimpo.presentation.screens.accountactivation.AccountActivationScreen
+import com.app.arcabyolimpo.presentation.screens.attendance.AttendanceListScreen
 import com.app.arcabyolimpo.presentation.screens.beneficiary.AddNewBeneficiaryScreen
 import com.app.arcabyolimpo.presentation.screens.beneficiary.BeneficiaryDetailScreen
 import com.app.arcabyolimpo.presentation.screens.beneficiary.BeneficiaryListScreen
-import com.app.arcabyolimpo.presentation.screens.capacitations.DisabilitiesList
+import com.app.arcabyolimpo.presentation.screens.beneficiary.ModifyBeneficiaryScreen
 import com.app.arcabyolimpo.presentation.screens.capacitations.DisabilitiesListScreen
 import com.app.arcabyolimpo.presentation.screens.capacitations.disabilitiesRegister.DisabilitiesRegisterScreen
+import com.app.arcabyolimpo.presentation.screens.capacitations.DisabilityDetailScreen
 import com.app.arcabyolimpo.presentation.screens.home.assistant.CollaboratorHomeScreen
 import com.app.arcabyolimpo.presentation.screens.home.coordinator.CoordinatorHomeScreen
 import com.app.arcabyolimpo.presentation.screens.home.scholar.ScholarHomeScreen
@@ -89,6 +92,10 @@ sealed class Screen(
         fun createRoute(userId: String) = "update_user/$userId"
     }
 
+    object AttendanceList : Screen("attendance_list/{userId}") {
+        fun createRoute(userId: String) = "attendance_list/$userId"
+    }
+
     object TokenVerification : Screen("user/verify-token?token={token}") {
         fun createRoute(token: String) = "user/verify-token?token=$token"
     }
@@ -133,9 +140,17 @@ sealed class Screen(
 
     object AddNewBeneficiary : Screen("beneficiary/create")
 
+    object ModifyBeneficiary : Screen("beneficiary/update/{beneficiaryId}"){
+        fun createRoute(beneficiaryId: String) = "beneficiary/update/$beneficiaryId"
+    }
+
     object CapacitationScreen : Screen("/disabilities/list")
 
     object DisabilitiesRegisterScreen : Screen("/disabilities/register")
+    
+    object DisabilityDetail : Screen("discapacity/{disabilityId}") {
+        fun createRoute(disabilityId: String) = "discapacity/$disabilityId"
+    }
 
     object SupplyDetail : Screen("supply/{idSupply}") {
         fun createRoute(idSupply: String) = "supply/$idSupply"
@@ -444,6 +459,9 @@ fun ArcaNavGraph(
                     navController.navigate(Screen.UpdateUserScreen.createRoute(id))
                 },
                 onDeleteClick = { navController.popBackStack() },
+                onAttendanceClick = { id ->
+                    navController.navigate(Screen.AttendanceList.createRoute(id))
+                },
             )
         }
 
@@ -475,6 +493,17 @@ fun ArcaNavGraph(
                 },
             )
         }
+
+        /** Attendance List Screen */
+        composable(
+            route = Screen.AttendanceList.route,
+            arguments = listOf(navArgument("userId") { type = NavType.StringType }),
+        ) {
+            AttendanceListScreen(
+                onBackClick = { navController.popBackStack() },
+            )
+        }
+
 
         /**
          * Workshops List Screen.
@@ -699,8 +728,29 @@ fun ArcaNavGraph(
         ) {
             BeneficiaryDetailScreen(
                 onBackClick = { navController.popBackStack() },
-                onModifyClick = { /* TODO: Lógica de VM */ },
+                onModifyClick = { beneficiaryId ->
+                    navController.navigate(Screen.ModifyBeneficiary.createRoute(beneficiaryId))
+                },
                 viewModel = hiltViewModel(),
+                beneficiaryId = it.arguments?.getString("beneficiaryId") ?: ""
+            )
+        }
+
+        /**
+         * Modify Beneficiary Screen.
+         *
+         * Allows the modification of an existing and active beneficiary.
+         */
+        composable(
+            route = Screen.ModifyBeneficiary.route,
+            arguments = listOf(navArgument("beneficiaryId") { type = NavType.StringType }),
+
+            ){ backStackEntry ->
+            val beneficiaryId = backStackEntry.arguments?.getString("beneficiaryId") ?: ""
+            ModifyBeneficiaryScreen(
+                navController = navController,
+                viewModel = hiltViewModel(),
+                beneficiaryId = beneficiaryId
             )
         }
 
@@ -714,8 +764,9 @@ fun ArcaNavGraph(
             DisabilitiesListScreen(
                 navController = navController,
                 onDisabilityClick = { id ->
+                    Log.d("Click", "Click")
                     // TODO: Navigate to disability detail when screen is created
-                    // navController.navigate(Screen.DisabilityDetail.createRoute(id))
+                    navController.navigate(Screen.DisabilityDetail.createRoute(id))
                 },
                 onAddClick = { navController.navigate(Screen.DisabilitiesRegisterScreen.route) },
                 onBackClick = { navController.popBackStack() },
@@ -726,6 +777,20 @@ fun ArcaNavGraph(
             DisabilitiesRegisterScreen(
                 onCreated = { navController.popBackStack() },
                 onBackClick = { navController.popBackStack() },
+            )
+        }
+
+        /**
+         * Disability Detail Screen.
+         *
+         * Shows the details of a disability.
+         */
+        composable(
+            route = Screen.DisabilityDetail.route
+        ) {
+            DisabilityDetailScreen(
+                onBackClick = { navController.popBackStack() },
+                viewModel = hiltViewModel(),
             )
         }