+++ app/src/main/java/com/app/arcabyolimpo/presentation/screens/capacitations/DisabilitiesList.kt
@@ -42,6 +42,7 @@ import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.navigation.NavHostController
 import com.app.arcabyolimpo.presentation.ui.components.atoms.alerts.Snackbarcustom
 import com.app.arcabyolimpo.presentation.ui.components.atoms.buttons.AddButton
+import com.app.arcabyolimpo.presentation.navigation.Screen
 import com.app.arcabyolimpo.presentation.ui.components.atoms.inputs.SearchInput
 import com.app.arcabyolimpo.presentation.ui.components.molecules.BeneficiaryCard
 import com.app.arcabyolimpo.ui.theme.ArcaByOlimpoTheme
@@ -103,6 +104,7 @@ fun DisabilitiesListScreen(
         onDisabilityClick = onDisabilityClick,
         onAddClick = onAddClick,
         onBackClick = onBackClick,
+        navController = navController
     )
 }
 
@@ -115,6 +117,7 @@ fun DisabilitiesList(
     onDisabilityClick: (String) -> Unit,
     onAddClick: () -> Unit,
     onBackClick: () -> Unit,
+    navController: NavHostController
 ) {
     ArcaByOlimpoTheme(darkTheme = true, dynamicColor = false) {
         Box(modifier = Modifier.fillMaxSize()) {
@@ -226,17 +229,16 @@ fun DisabilitiesList(
                                 items(state.disabilities, key = { it.id }) { disability ->
                                     BeneficiaryCard(
                                         name = disability.name,
-                                        onClick = { onDisabilityClick(disability.id) },
-                                        cardModifier =
-                                            Modifier
-                                                .fillMaxWidth()
-                                                .padding(horizontal = 8.dp)
-                                                .height(160.dp),
-                                        contentPadding =
-                                            PaddingValues(
-                                                vertical = 20.dp,
-                                                horizontal = 40.dp,
-                                            ),
+                                        onClick = {
+                                            navController.navigate(Screen.DisabilityDetail.createRoute(disability.id))},
+                                        cardModifier = Modifier
+                                            .fillMaxWidth()
+                                            .padding(horizontal = 8.dp)
+                                            .height(160.dp),
+                                        contentPadding = PaddingValues(
+                                            vertical = 20.dp,
+                                            horizontal = 40.dp,
+                                        ),
                                     )
                                 }
                             }