+++ CHANGELOG.md
@@ -1,6 +1,19 @@
+## [0.2.0] - 02/12/2025
+### Added
+- Consult disability details
+  
+## [0.2.0] - 01/12/2025
+### Changed
+ - Modify topbar for scaffold
+
 ## [0.2.0] - 01/12/2025
 ### Added
+- Cache for product list and detail
+- Modify beneficiary
 - Register Disability Functionalities
+
+### Fixed
+- Debug modify beneficiary
 
 ## [0.2.0] - 30/11/2025
 ### Added
@@ -9,10 +22,15 @@
 ### Changed
 - Added a wey to prevent certain chars and emojis in workshop inputs. 
 
+### Fixed
+- Fixed attendance date/time formatting.
+- Improved "View Attendance" button alignment and styling.
+
 ## [0.2.0] - 29/11/2025
 ### Added
 - Added offline for product batch list and detail
 - Added a Capacitations view for Disabilities List
+- Added offline for supply batches
 
 ## [0.2.0] - 29/11/2025
 
@@ -24,14 +42,15 @@
 - Change beneficiary detail for showing more disabilities
 - Change so now Beneficiaries can be registered with more disabilities
 
+## [0.1.0] - 25/11/2025
+### Added
+- US for consult attendance
+
 ## [0.1.0] - 20/11/2025
 ### Added
 - ModifyProduct validation fields
 - RegisterProduct validation fields
 - Arrow to return register and modify product
-
-## [0.1.0] - 20/11/2025
-### Added
 - US for filter and order workshops.
 - Functional go back button to Supplies List Screen
 - Functional go back buttons to Orders and Inventory Screens